Restructure to unified folder
