--- cpu/fe310/include/cpu.h
@@ -1,29 +1,0 @@
-/*
- * Copyright (C) 2023 Gunar Schorcht
- *
- * This file is subject to the terms and conditions of the GNU Lesser
- * General Public License v2.1. See the file LICENSE in the top level
- * directory for more details.
- */
-
-#pragma once
-
-/**
- * @ingroup     cpu_fe310
- * @{
- *
- * @file
- * @brief       CPU specific definitions
- */
-
-#include "cpu_common.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-#ifdef __cplusplus
-}
-#endif
-
-/** @} */
