+++ cpu/gd32v/include/cpu_conf.h
@@ -73,6 +73,23 @@ typedef enum {
 #define CONFIG_AFIO_PCF0_SWJ_CFG    SWJ_CFG_NO_NJTRST
 #endif
 
+/**
+ * @brief   Returns the address of running application in flash
+ */
+static inline uint32_t cpu_get_image_baseaddr(void)
+{
+    extern uint8_t _start;
+    return (uint32_t)&_start;
+}
+
+/**
+ * @brief   Starts another image in flash
+ */
+static inline void cpu_jump_to_image(uint32_t addr)
+{
+    __asm__ volatile ("jr %0" :: "r" (addr));
+}
+
 #ifdef __cplusplus
 }
 #endif
