+++ cpu/rp2350/Kconfig
@@ -5,6 +5,7 @@
 config CPU_FAM_RP2350
     bool
     select CPU_CORE_CORTEX_M33
+    select CPU_CORE_RV32IMAC
 
 config CPU_FAM
     default "RP2350" if CPU_FAM_RP2350
@@ -18,6 +19,3 @@ config CPU_MODEL
 
 config CPU
     default "rp2350" if CPU_FAM_RP2350
-
-
-source "$(RIOTCPU)/cortexm_common/Kconfig"
