+++ cpu/rp2350/Makefile.dep
@@ -1,3 +1,7 @@
 USEPKG += picosdk
 
-include $(RIOTCPU)/cortexm_common/Makefile.dep
+ifeq (,$(filter rp2350_arm,$(USEMODULE)))
+		DIRS += $(RIOTCPU)/cortexm_common
+else
+		DIRS += $(RIOTCPU)/riscv_common
+endif
