+++ cpu/rp2350/Makefile.features
@@ -1,9 +1,18 @@
-CPU_CORE    := cortex-m33
-CPU_FAM     := RP2350
-CPU_ARCH    = armv8m
-CPU_MODEL   = rp2350
+ifeq (,$(filter rp2350_arm,$(USEMODULE)))
+		CPU_CORE    := cortex-m33
+		CPU_FAM     := RP2350
+		CPU_ARCH    = armv8m
+		CPU_MODEL   = rp2350
 
-include $(RIOTCPU)/cortexm_common/Makefile.features
+		include $(RIOTCPU)/cortexm_common/Makefile.features
+else
+		CPU_CORE    := rv32
+		CPU_FAM     := RP2350_RISCV
+		CPU_ARCH    = rv32
+		CPU_MODEL   = rp2350_riscv
+
+		include $(RIOTCPU)/riscv_common/Makefile.features
+endif
 
 FEATURES_PROVIDED += periph_gpio
 FEATURES_PROVIDED += periph_uart
