+++ cpu/rp2350/Makefile.include
@@ -20,14 +20,24 @@ LINKFLAGS += -mcpu=$(CPU_ARCH) -mthumb
 LINKFLAGS += -Wl,--gc-sections
 LINKFLAGS += -Wl,--start-group -lc -lm -Wl,--end-group
 
-# Vector table configuration
-VECTORS_O       ?= $(BINDIR)/cpu/vectors.o
-VECTORS_FILE    := $(RIOTCPU)/rp2350/vectors.c
-
 # Supported programmers and debuggers
 PROGRAMMERS_SUPPORTED := picotool openocd jlink
 PROGRAMMER ?= picotool
 OPENOCD_DEBUG_ADAPTER ?= dap
 
-# Include the base Cortex-M makefile
-include $(RIOTMAKE)/arch/cortexm.inc.mk
+ifeq (,$(filter rp2350_arm,$(USEMODULE)))
+		include $(RIOTCPU)/cortexm_common/Makefile.features
+		# Vector table configuration
+		VECTORS_O       ?= $(BINDIR)/cpu/vectors.o
+		VECTORS_FILE    := $(RIOTCPU)/rp2350/vectors.c
+
+		# Include the base Cortex-M makefile
+		include $(RIOTMAKE)/arch/cortexm.inc.mk
+else
+		CFLAGS += -march=rv32imac_zicsr_zifencei_zba_zbb_zbkb_zbs
+		CFLAGS += -mabi=lp64
+		LINKFLAGS += -march=rv32imac_zicsr_zifencei_zba_zbb_zbkb_zbs
+		LINKER_SCRIPT ?= riscv.ld
+
+		include $(RIOTCPU)/riscv_common/Makefile.include
+endif
